+++ client/src/App.js
@@ -4,6 +4,7 @@ import HomePage from './pages/HomePage';
 import SampleFormPage from './pages/SampleFormPage';
 import MapToolPage from './pages/MapToolPage';
 import ApiTestPage from './pages/ApiTestPage';
+import SurveyListPage from './pages/SurveyListPage';
 
 import './App.css';
 
@@ -16,6 +17,7 @@ class App extends Component {
           <Route path="/sample_form" component={SampleFormPage}/>
           <Route path="/map_tool" component={MapToolPage}/>
           <Route path="/api_test" component={ApiTestPage}/>
+          <Route path="/surveys" component={SurveyListPage}/>
         </Switch>
       </div>
     );