+++ client/src/services/api.js
@@ -35,3 +35,14 @@ export const sendSql = async (args) => {
 
   return response.json();
 }
+
+export const getSurveyDefinitions = async (args) => {
+  const response = await fetch('/api/surveys');
+  const body = await response.json();
+
+  console.log(body);
+  
+  if (response.status !== 200) throw Error(body.message);
+
+  return body;
+}